requirements.test.txt
pytest
pytest-homeassistant-custom-component
pytest-asyncio